+++ engine/src/flutter/lib/web_ui/lib/src/engine/renderer.dart
@@ -249,7 +249,7 @@ abstract class Renderer {
   ui.ParagraphStyle createParagraphStyle({
     ui.TextAlign? textAlign,
     ui.TextDirection? textDirection,
-    double? maxLines,
+    int? maxLines,
     String? fontFamily,
     double? fontSize,
     double? height,
