+++ engine/src/flutter/lib/web_ui/test/ui/text_golden_test.dart
@@ -637,7 +637,7 @@ Future<void> testTextStyle(
   // ParagraphStyle properties
   ui.TextAlign? paragraphTextAlign,
   ui.TextDirection? paragraphTextDirection,
-  double? paragraphMaxLines,
+  int? paragraphMaxLines,
   String? paragraphFontFamily,
   double? paragraphFontSize,
   double? paragraphHeight,
